remove powerline by default
this is obnoxious; i don't want people to have to install powerline
fonts. it's not so bad on Linux since you can install it separately and
it'll fallback to your normally configured font, but on OS X you're
limited to fonts that have been spliced with the powerline symbols.

if you want to re-enable this, place the following in ~/.nvimrc.local:

    let g:lightline.component.lineinfo = '%3l:%-2v'
    let g:lightline.separator    = { 'left': '', 'right': '' }
    let g:lightline.subseparator = { 'left': '', 'right': '' }

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>